requirements.txt
# Python dependencies for AI Video Platform

# Background removal for stickers (IMG-003)
rembg>=2.0.50
pillow>=10.0.0

# Image processing
numpy>=1.24.0
